nominal changes, test for github email contact info

+++ NWFC Reserve/EQRScheduleTopVCntrllr.m
@@ -1477,7 +1477,7 @@
             letterString = @"Su";
         }
         
-        NSString *dateString = [NSString stringWithFormat:@"%ld", indexPath.row + 1];
+        NSString *dateString = [NSString stringWithFormat:@"%d", (NSInteger)indexPath.row + 1];
         
         //delete the datestring if the month doesn't extend that far
         if (!letterString){
